+++ annotation_api/annotation_api.go
@@ -254,9 +254,9 @@ func init() {
 
 		http := NewHttp(zkCon)
 
-		manager.RegisterHttpEndpoint("DELETE", httpPrefix+"/:serviceId", http.deleteAnnotation)
-		manager.RegisterHttpEndpoint("GET", httpPrefix, http.listAnnotations)
-		manager.RegisterHttpEndpoint("OPTIONS", httpPrefix+"/:serviceId", http.optionsAnnotation)
-		manager.RegisterHttpEndpoint("POST", httpPrefix+"/:serviceId", http.createAnnotation)
+		manager.RegisterHttpHandleFunc("DELETE", httpPrefix+"/:serviceId", http.deleteAnnotation)
+		manager.RegisterHttpHandleFunc("GET", httpPrefix, http.listAnnotations)
+		manager.RegisterHttpHandleFunc("OPTIONS", httpPrefix+"/:serviceId", http.optionsAnnotation)
+		manager.RegisterHttpHandleFunc("POST", httpPrefix+"/:serviceId", http.createAnnotation)
 	}
 }